refactor: term(startAt, endAt) 객체 분리
