+++ backend/src/test/java/com/staccato/memory/domain/MemoryTest.java
@@ -1,8 +1,6 @@
 package com.staccato.memory.domain;
 
-import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.AssertionsForClassTypes.assertThatCode;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
@@ -17,24 +15,6 @@ import com.staccato.fixture.moment.MomentFixture;
 import com.staccato.moment.domain.Moment;
 
 class MemoryTest {
-    @DisplayName("끝 날짜는 시작 날짜보다 앞설 수 없다.")
-    @Test
-    void validateDate() {
-        assertThatCode(() -> MemoryFixture.create(LocalDate.now().plusDays(1), LocalDate.now()))
-                .isInstanceOf(StaccatoException.class)
-                .hasMessage("끝 날짜가 시작 날짜보다 앞설 수 없어요.");
-    }
-
-    @DisplayName("특정 날짜가 추억 날짜에 속하는지 알 수 있다.")
-    @Test
-    void withoutDuration() {
-        // given
-        Memory memory = MemoryFixture.create(LocalDate.of(2023, 7, 1), LocalDate.of(2023, 7, 10));
-
-        // when & then
-        assertThat(memory.isWithoutDuration(LocalDateTime.of(2023, 7, 11, 10, 0))).isTrue();
-    }
-
     @DisplayName("추억을 수정 시 기존 순간 기록 날짜를 포함하지 않는 경우 수정에 실패한다.")
     @Test
     void validateDuration() {
